feat: ship self-contained operations skill
