+++ .github/workflows/release.yml
@@ -173,10 +173,28 @@ jobs:
           path: dist/*
           if-no-files-found: error
 
+  build-ops:
+    name: Build operations bundle
+    needs: verify
+    runs-on: ubuntu-22.04
+    steps:
+      - name: Check out source
+        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
+
+      - name: Build platform-independent operations archive
+        run: scripts/package-ops.sh
+
+      - name: Upload workflow artifact
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7
+        with:
+          name: masque-ops
+          path: dist/*
+          if-no-files-found: error
+
   publish:
     name: Publish GitHub Release
     if: startsWith(github.ref, 'refs/tags/v')
-    needs: [build, build-probe-macos]
+    needs: [build, build-probe-macos, build-ops]
     runs-on: ubuntu-22.04
     permissions:
       contents: write
