+++ Cargo.toml
@@ -10,7 +10,7 @@ undocumented_unsafe_blocks = "deny"
 
 [package]
 name = "masque-server"
-version = "0.12.2"
+version = "0.13.0"
 edition = "2024"
 rust-version = "1.88"
 description = "High-performance MASQUE proxy for CONNECT, CONNECT-UDP, and CONNECT-IP over HTTP/2 and HTTP/3"
