+++ docs/operations.md
@@ -1,10 +1,10 @@
 # AI-assisted fleet operations
 
-The repository includes an optional operations layer for repeatable MASQUE
-fleet installation and upgrades. An AI coding agent can interpret an operator's request through
-the repo-scoped `masque-ops` Skill, but all privileged work is performed by the
-deterministic, tested `scripts/masque-ops.py` CLI. No model, API key, or AI
-runtime is installed on a proxy server.
+MASQUE Server provides an optional operations layer for repeatable fleet
+installation and upgrades. An AI coding agent can interpret an operator's
+request through the installable `masque-ops` Skill, but all privileged work is
+performed by its deterministic, tested CLI. No model, API key, or AI runtime is
+installed on a proxy server.
 
 This is deliberately bounded release operations rather than general configuration
 management. It can bootstrap an explicitly described empty host, inspect status,
@@ -15,14 +15,40 @@ certificates, users, routes, forwarding, firewall, NAT, or monitoring services.
 It targets the paths and service name used by the official Linux installer;
 custom installation layouts are intentionally outside its scope.
 
-## Set up an inventory
+## Install the operations bundle
 
 Python 3.11 or newer and OpenSSH are required on the administration machine.
-Copy the example outside the repository and protect it:
+Linux and macOS can install the Skill, CLI, inventory example, and matching
+local probe from the latest stable Release with:
+
+```sh
+curl -fsSL https://raw.githubusercontent.com/Vincent-bin/masque-server/main/install-ops.sh | sh
+```
+
+The installer verifies the platform-independent operations archive and the
+platform-specific probe archive against their Release SHA-256 sidecars before
+activation. It installs the Skill at `~/.agents/skills/masque-ops`, the CLI and
+probe under `~/.local/bin` for an unprivileged user, and the example under
+`~/.config/masque-server`. Re-running it upgrades these release-managed files
+without reading or changing private inventory, state, or credentials. Pin an
+exact version with:
+
+```sh
+curl -fsSL https://raw.githubusercontent.com/Vincent-bin/masque-server/main/install-ops.sh | \
+  env MASQUE_VERSION=vX.Y.Z sh
+```
+
+Set `MASQUE_OPS_INSTALL_PROBE=0` to skip the probe. A repository checkout also
+works through `scripts/masque-ops.py`, and its repo-scoped Skill is discovered
+without running this installer.
+
+## Set up an inventory
+
+Copy the installed example and protect the private file:
 
 ```sh
 install -d -m 0700 ~/.config/masque-server ~/.config/masque-server/secrets
-install -m 0600 deploy/config/fleet.example.toml \
+install -m 0600 ~/.config/masque-server/fleet.example.toml \
   ~/.config/masque-server/fleet.toml
 ```
 
@@ -40,8 +66,8 @@ addresses, key paths, remote TLS paths, usernames, passwords, and generated
 client-file paths. The deterministic CLI is the only component that parses the
 private inventory or handles secret bytes.
 
-Install the release-matched probe on Linux or macOS x86_64/ARM64 with automatic
-platform and architecture detection:
+If it was omitted from the operations install, install the release-matched
+probe separately on Linux or macOS x86_64/ARM64:
 
 ```sh
 curl -fsSL https://raw.githubusercontent.com/Vincent-bin/masque-server/main/install-probe.sh | sh
@@ -60,10 +86,10 @@ forwarding, local commands, and shared control sockets.
 ## Read-only inspection
 
 ```sh
-scripts/masque-ops.py validate
-scripts/masque-ops.py status
-scripts/masque-ops.py plan --version vX.Y.Z
-scripts/masque-ops.py diagnose edge-a --journal-lines 100
+masque-ops validate
+masque-ops status
+masque-ops plan --version vX.Y.Z
+masque-ops diagnose edge-a --journal-lines 100
 ```
 
 `status` checks the installed version, architecture, service state,
@@ -82,10 +108,10 @@ sharing it. Place `--json` before the subcommand for structured output.
 Mutations require an exact version and the explicit `--apply` gate:
 
 ```sh
-scripts/masque-ops.py plan --version vX.Y.Z
-scripts/masque-ops.py deploy edge-a --version vX.Y.Z --apply
-scripts/masque-ops.py rollout --version vX.Y.Z --apply
-scripts/masque-ops.py rollback edge-a --apply
+masque-ops plan --version vX.Y.Z
+masque-ops deploy edge-a --version vX.Y.Z --apply
+masque-ops rollout --version vX.Y.Z --apply
+masque-ops rollback edge-a --apply
 ```
 
 `rollout` requires exactly one selected canary and normally requires an
@@ -100,20 +126,20 @@ possible, and the rollout stops without touching later hosts.
 
 Add `[hosts.bootstrap]` to the host's private inventory entry. It describes the
 authentication mode, ports, existing remote ACME certificate paths, and local
-mode-`0600` destinations for generated credentials. The repository example
+mode-`0600` destinations for generated credentials. The installed example
 shows Basic mode; `client_cert` and `dual` additionally require
 `client_endpoint` and `client_config`. Inline passwords are rejected.
 
 First run the non-mutating preflight:
 
 ```sh
-scripts/masque-ops.py bootstrap edge-new --version vX.Y.Z
+masque-ops bootstrap edge-new --version vX.Y.Z
 ```
 
 After explicit approval, the same command performs the install:
 
 ```sh
-scripts/masque-ops.py bootstrap edge-new --version vX.Y.Z --apply
+masque-ops bootstrap edge-new --version vX.Y.Z --apply
 ```
 
 The CLI requires a verified root SSH entry for this one operation. It confirms
@@ -147,7 +173,7 @@ and relies on the package installer for atomic replacement and service-state
 rollback.
 
 This narrow command can be granted to a dedicated SSH account with sudo. See
-the repo Skill's
+the Skill's
 [`inventory.md`](../.agents/skills/masque-ops/references/inventory.md) for the
 sudoers example and probe configuration. A server running an older release
 does not yet have the helper; its first transition requires the explicit,
@@ -156,9 +182,9 @@ that transition.
 
 ## Using the Skill
 
-Codex-compatible agents discover the repo-scoped
-`.agents/skills/masque-ops/SKILL.md` automatically when working in this
-repository. A typical request is: “Use `$masque-ops` to inspect the fleet and
+Codex-compatible agents discover either the installed
+`~/.agents/skills/masque-ops/SKILL.md` or the repo-scoped copy automatically
+after startup. A typical request is: “Use `$masque-ops` to inspect the fleet and
 prepare an upgrade plan for `vX.Y.Z`.” Inspection and planning do not authorize
 deployment; ask explicitly to execute the reviewed plan before the agent may
 add `--apply`.
