+++ tools/masque-probe/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "masque-probe"
-version = "0.12.2"
+version = "0.13.0"
 edition = "2024"
 rust-version = "1.88"
 description = "End-to-end connectivity diagnostics for MASQUE proxies"
